+++ changes.md
@@ -10,3 +10,4 @@ Changes that I've added on top of MarcoDBAA's source:
 - Add parameter for the "repaint strategy". Defaults to “repaint if any of the app window is in focus”.
 - Limit FPS to the target FPS when game is repainted. The FPS is unlimited when the world is not repainted (e.g. in the background).
 - Save world as png when doing auto backup
+- Ignore auto backup if the world was not saved yet. This is to avoid stopping the simulation and waiting for the user to choose a file.